+++ .bashrc
@@ -1,0 +1,2 @@
+export PATH="C:\Users\Moturu\anaconda\Scripts"
+alias python="winpty python.exe"
