+++ setup.cfg
@@ -1,0 +1,4 @@
+[metadata]
+# This includes the license file(s).
+
+license_files = LICENSE.txt
